added in info_file header and cpp files to offload info file specific constants, structs, and functions from control
