+++ src/cxx_tools/file_utility.h
@@ -25,6 +25,10 @@ const std::string DEFAULT_DATE_FORMAT = "(%m_%d_%Y)";
 const std::string DEFAULT_TIME_FORMAT = "(%H:%M:%S)";
 const std::string DEFAULT_LOCALE_FORMAT = "(%z %Z)";
 
+const std::string TXT_EXT = ".txt";
+const std::string BIN_EXT = ".bin";
+const std::string SIM_EXT = ".sim";
+
 std::string get_file_basename(std::string full_file_path);
 
 std::string get_current_time_as_string();
